a LOT of not interesting stuff

+++ main.py
@@ -1,39 +1,45 @@
 import sys
-import config
 import random
+import hashlib
 import discord
 from gtts import gTTS
 import pymysql.cursors
+from config import config
 from colorama import Fore
-from debug import dbprint
+from bloodyterminal import btext
 from discord.ext import commands
-from TwitterAPI import TwitterAPI
 
-api = TwitterAPI(config.appKey, config.appSecret, config.oauthToken, config.oauthTokenSecret)
 description = '''A learning experience bot, written in python using discory.py .'''
 bot = commands.Bot(command_prefix='.', description=description)
 
 
 @bot.event
 async def on_ready():
-    dbprint('success', 'Logged in as')
-    dbprint('info', "Name: " + Fore.GREEN + bot.user.name)
-    dbprint('info', "ID: " + Fore.GREEN + bot.user.id)
-    dbprint('info', '________________________')
-    dbprint('info', sys.platform)
+    btext.info('Logged in as')
+    btext.info('Name: ' + Fore.GREEN + bot.user.name)
+    btext.info("ID: " + Fore.GREEN + bot.user.id)
+    btext.info('________________________')
+    btext.info(sys.platform)
     if sys.platform != "win32":
-        await bot.change_presence(game=discord.Game(name='.info'))
+        await bot.change_presence(game=discord.Game(name='.info', url="https://twitch.tv/TheBloodyScreen", type=1))
+        btext.success("game set successfully")
     else:
-        await bot.change_presence(game=discord.Game(name='.info - indev'))
-    dbprint('success', "connected successfully")
-    dbprint('info', '________________________')
+        await bot.change_presence(game=discord.Game(name='.info - indev', url="https://twitch.tv/TheBloodyScreen", type=1))
+        btext.success("game set successfully")
+    btext.success("connected successfully")
+    btext.info('________________________')
 
 
 @bot.command(pass_context=True)
 async def info(ctx):
     await bot.delete_message(ctx.message)
     await bot.send_typing(ctx.message.author)
-    await bot.send_message(ctx.message.author, embed=discord.Embed(title='Commandlist for TheBloodyDiscord.', description="\n!info\nCauses these messages.\n\n!faq\nPosts a link to the FAQ in the channel it was used in.\n\n!twitch\nPosts a link to the twitch channel in the channel it was used in.\n\nIf the bot says 'Playing .info - indev' I am currently working on the bot and there might be problems while using commands.", color=0x2AE92A))
+    await bot.send_message(ctx.message.author, embed=discord.Embed(title='list of commands for TheBloodyDiscord.',
+                                                                   description="\n!info\nCauses these messages."
+                                                                               "\n\n!faq\nPosts a link to the FAQ in the channel it was used in."
+                                                                               "\n\n!twitch\nPosts a link to the twitch channel in the channel it was used in."
+                                                                               "\n\nIf the bot says 'Playing .info - indev' I am currently working on the bot and there might be problems while using commands.",
+                                                                   color=0x2AE92A))
 
 
 @bot.command(pass_context=True)
@@ -80,8 +86,8 @@ async def play(ctx, cmd: str):
 @voice.command(pass_context=True)
 async def tts(ctx, cmd: str):
     if cmd:
-        tts = gTTS(text=str(ctx.message.content).replace('.voice tts ', ''), lang='en')
-        tts.save('./audio/tts.mp3')
+        ttos = gTTS(text=str(ctx.message.content).replace('.voice tts ', ''), lang='en')
+        ttos.save('./audio/tts.mp3')
         global voice
         voice = await bot.join_voice_channel(ctx.message.author.voice_channel)
         global player
@@ -103,6 +109,20 @@ async def rimshot(ctx):
     voice = await bot.join_voice_channel(ctx.message.author.voice_channel)
     global player
     player = voice.create_ffmpeg_player('./audio/rimshot.mp3')
+    player.start()
+    check = True
+    while check is True:
+        if player.is_done():
+            await voice.disconnect()
+            check = False
+
+
+@voice.command(pass_context=True)
+async def wirsindgeil(ctx):
+    global voice
+    voice = await bot.join_voice_channel(ctx.message.author.voice_channel)
+    global player
+    player = voice.create_ffmpeg_player('./audio/music.mp3')
     player.start()
     check = True
     while check is True:
@@ -137,17 +157,6 @@ async def leave():
 # everything voice --end
 
 
-# simple twitter post command
-@bot.command(pass_context=True)
-async def tweet(ctx):
-    await bot.delete_message(ctx.message)
-    if str(ctx.message.author) == 'TheBloodyScreen#4278':
-        r = api.request('statuses/update', {'status': ctx.message.content.replace('.tweet ', '')})
-        await bot.send_message(ctx.message.author, 'SUCCESS' if r.status_code == 200 else 'FAILURE')
-    else:
-        await bot.say("You don't have permission to use this command!")
-
-
 # simple addition
 @bot.command()
 async def add(first: int, second: int):
@@ -171,7 +180,7 @@ async def repeat(ctx, times: int, *, content: str):
         await bot.say("You don't have permission to use this command!")
 
 
-# It's like "the purge," Morty. That movie "the purge"?
+# It's like "the purge," Morty. You know that movie "the purge"?
 @bot.command(pass_context=True)
 async def purge(ctx, amount):
     if 'admins' in getRoles(ctx.message.author):
@@ -193,14 +202,19 @@ async def roles(ctx):
         for member in server.members:
             if member.mentioned_in(ctx.message):
                 user = member
-    await bot.send_message(ctx.message.channel, embed=discord.Embed(title=str(str(user)[0:-5]) + ' has the following Groups:', description=("\n".join(getRoles(user))), color=0x0000FF))
+                await bot.send_message(ctx.message.channel, embed=discord.Embed(title=str(str(user)[0:-5]) + ' has the following Groups:', description=("\n".join(getRoles(user))), color=0x0000FF))
 
 
 # tell me something fun
 @bot.command(pass_context=True)
 async def quote(ctx):
     await bot.delete_message(ctx.message)
-    connection = pymysql.connect(host=config.host, user=config.user, password=config.password, db=config.db, charset=config.charset, cursorclass=pymysql.cursors.DictCursor)
+    connection = pymysql.connect(host=config['database']['discord']['host'],
+                                 user=config['database']['discord']['user'],
+                                 password=config['database']['discord']['pass'],
+                                 db=config['database']['discord']['db'],
+                                 charset=config['database']['discord']['charset'],
+                                 cursorclass=pymysql.cursors.DictCursor)
 
     try:
         with connection.cursor() as cursor:
@@ -210,8 +224,26 @@ async def quote(ctx):
             sql = "SELECT `quote` FROM `quotes` WHERE `id`=%s"
             cursor.execute(sql, (random.randint(1, int(maxid))))
             result = (str(cursor.fetchone()).replace("{'quote': ", '')).replace('}', '')
-            dbprint('info', result)
             await bot.say(result)
+    finally:
+        connection.close()
+
+
+@bot.command(pass_context=True)
+async def qamount(ctx):
+    await bot.delete_message(ctx.message)
+    connection = pymysql.connect(host=config['database']['discord']['host'],
+                                 user=config['database']['discord']['user'],
+                                 password=config['database']['discord']['pass'],
+                                 db=config['database']['discord']['db'],
+                                 charset=config['database']['discord']['charset'],
+                                 cursorclass=pymysql.cursors.DictCursor)
+    try:
+        with connection.cursor() as cursor:
+            sql = "SELECT TABLE_NAME AS 'Table Name', TABLE_ROWS AS 'Rows' FROM information_schema.TABLES WHERE TABLES.TABLE_SCHEMA = 'TheBloodyDiscord' AND TABLES.TABLE_TYPE = 'BASE TABLE'"
+            cursor.execute(sql)
+            maxid = (str(cursor.fetchall()[0]).replace("{'Table Name': 'quotes', 'Rows': ", '')).replace('}', '')
+            await bot.say(str(maxid))
     finally:
         connection.close()
 
@@ -219,7 +251,12 @@ async def quote(ctx):
 @bot.command(pass_context=True)
 async def quoteid(ctx):
     await bot.delete_message(ctx.message)
-    connection = pymysql.connect(host=config.host, user=config.user, password=config.password, db=config.db, charset=config.charset, cursorclass=pymysql.cursors.DictCursor)
+    connection = pymysql.connect(host=config['database']['discord']['host'],
+                                 user=config['database']['discord']['user'],
+                                 password=config['database']['discord']['pass'],
+                                 db=config['database']['discord']['db'],
+                                 charset=config['database']['discord']['charset'],
+                                 cursorclass=pymysql.cursors.DictCursor)
 
     try:
         with connection.cursor() as cursor:
@@ -234,7 +271,12 @@ async def quoteid(ctx):
 @bot.command(pass_context=True)
 async def addquote(ctx):
     await bot.delete_message(ctx.message)
-    connection = pymysql.connect(host=config.host, user=config.user, password=config.password, db=config.db, charset=config.charset, cursorclass=pymysql.cursors.DictCursor)
+    connection = pymysql.connect(host=config['database']['discord']['host'],
+                                 user=config['database']['discord']['user'],
+                                 password=config['database']['discord']['pass'],
+                                 db=config['database']['discord']['db'],
+                                 charset=config['database']['discord']['charset'],
+                                 cursorclass=pymysql.cursors.DictCursor)
 
     try:
         with connection.cursor() as cursor:
@@ -246,13 +288,91 @@ async def addquote(ctx):
         connection.close()
 
 
+# wow account stuff
+@bot.group(pass_context=True)
+async def account(ctx):
+    channel = ctx.message.channel
+
+    if ctx.invoked_subcommand is None:
+        if str(channel.type) != 'private':
+            await bot.say('invalid link or subcommand')
+            await bot.say('Also: these commands only work in private chats with me')
+
+
+@account.command(pass_context=True)
+async def create(ctx):
+    connection = pymysql.connect(host=config['database']['trinity']['host'],
+                                 user=config['database']['trinity']['user'],
+                                 password=config['database']['trinity']['pass'],
+                                 db=config['database']['trinity']['db'],
+                                 cursorclass=pymysql.cursors.DictCursor)
+    message = ctx.message.content.split()
+    hexhash = hashlib.sha1((message[2] + ':' + message[3]).upper().encode('utf-8')).hexdigest()
+    if "--" in ctx.message.content:
+        await bot.say('You used illegal characters. Please try again!')
+        pass
+    else:
+        try:
+            with connection.cursor() as cursor:
+                sql = "INSERT INTO `account` (`username`, `sha_pass_hash`, `expansion`) VALUES (%s, %s, %s)"
+                cursor.execute(sql, (message[2].upper(), hexhash.upper(), 2))
+            connection.commit()
+            await bot.say("account created! you should now be able to login to the server")
+            await bot.say("remember to set the realmlist to thebloodyscreen.com")
+        except:
+            await bot.say("the account creation seems to have failed, please contact TheBloodyScreen!")
+        finally:
+            connection.close()
+
+
+@account.command(pass_context=True)
+async def delete(ctx):
+    connection = pymysql.connect(host=config['database']['trinity']['host'],
+                                 user=config['database']['trinity']['user'],
+                                 password=config['database']['trinity']['pass'],
+                                 db=config['database']['trinity']['db'],
+                                 cursorclass=pymysql.cursors.DictCursor)
+    message = ctx.message.content.split()
+
+    if "--" in ctx.message.content:
+        await bot.say('You used illegal characters. Please try again!')
+        pass
+    else:
+        try:
+            with connection.cursor() as cursor:
+                sql = "DELETE FROM `account` WHERE username = %s AND sha_pass_hash = %s"
+                btext.debug(hashlib.sha1((message[2] + ':' + message[3]).upper().encode('utf-8')).hexdigest().upper())
+                cursor.execute(sql, (message[2].upper(), hashlib.sha1((message[2] + ':' + message[3]).upper().encode('utf-8')).hexdigest().upper()))
+            connection.commit()
+            await bot.say("account deleted! THIS IS NOT REVERSABLE!")
+        except pymysql.Error as e:
+            btext.debug(str(e))
+            await bot.say("the account deletion seems to have failed, please contact TheBloodyScreen!")
+        finally:
+            connection.close()
+
+
+@account.command(pass_context=True)
+async def help(ctx):
+    channel = ctx.message.channel
+    if str(channel.type) == 'private':
+        await bot.send_message(ctx.message.author, embed=discord.Embed(title='TheBloodyLichKing Help', description="""\nThe available subcommands are:
+                                                                                                                      \n\nhelp\ndisplays this message
+                                                                                                                      \n\ncreate name password\ntries to create server account: 'name' with password 'password'
+                                                                                                                      \n\ndelete name password\ntries to delete server account: 'name' with password 'password(NOT YET IMPLEMENTED)'
+                                                                                                                      \n\nExample:
+                                                                                                                      \n.account create testname testpassword""", color=0x2900AD))
+    else:
+        await bot.say('DO NOT use these commands in public channels! Please PM me!')
+
+
 # markdown helper
 @bot.event
 async def on_message(message):
     author = message.author
     bot.get_all_emojis()
 
-    if message.content.startswith('CODE') and message.author != 'TheBloodyDiscord#6405' and message.author.id != '322081774622605312':
+    if message.content.startswith('CODE') and message.author != 'TheBloodyDiscord#6405' and message.author.id != '322081774622605312' and message.author.id != '385007074863480832':
         author = str(message.author)
         await bot.delete_message(message)
         newMessage = str(message.content).replace("CODE", "```")
@@ -265,9 +385,13 @@ async def on_message(message):
         await bot.add_reaction(message, ":vagina:234426457068273664")
     elif message.content.find('boob') >= 0:
         await bot.add_reaction(message, ":titcart:234430485437218829")
+        await bot.send_message(message.channel, "<:titcart:234430485437218829>")
     elif message.content.find('ALLEMOJIS') >= 0:
         for emoji in bot.get_all_emojis():
             await bot.add_reaction(message, emoji)
+    elif message.content.startswith("<@322081774622605312>"):
+        await bot.send_message(message.channel, "\n%s\nI'm having a hard time replying to mentions.\nPlease use !info." % author.mention)
+
     await bot.process_commands(message)
 
 
@@ -278,4 +402,4 @@ def getRoles(author):
     return roles
 
 
-bot.run(config.token)
+bot.run(config['discord']['token'])
